Move ManageViewModels from Models to ViewModels namespace.
