--- src/KPcore/Models/ManageViewModels/FactorViewModel.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace KPcore.Models.ManageViewModels
-{
-    public class FactorViewModel
-    {
-        public string Purpose { get; set; }
-    }
-}
