+++ src/KPcore/ViewModels/ManageViewModels/FactorViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace KPcore.ViewModels.ManageViewModels
+{
+    public class FactorViewModel
+    {
+        public string Purpose { get; set; }
+    }
+}
