+++ src/KPcore/ViewModels/ManageViewModels/IndexViewModel.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 using Microsoft.AspNetCore.Identity;
 
-namespace KPcore.Models.ManageViewModels
+namespace KPcore.ViewModels.ManageViewModels
 {
     public class IndexViewModel
     {
