+++ src/KPcore/ViewModels/ManageViewModels/ManageLoginsViewModel.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 using Microsoft.AspNetCore.Http.Authentication;
 using Microsoft.AspNetCore.Identity;
 
-namespace KPcore.Models.ManageViewModels
+namespace KPcore.ViewModels.ManageViewModels
 {
     public class ManageLoginsViewModel
     {
