+++ src/KPcore/ViewModels/ManageViewModels/RemoveLoginViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace KPcore.ViewModels.ManageViewModels
+{
+    public class RemoveLoginViewModel
+    {
+        public string LoginProvider { get; set; }
+        public string ProviderKey { get; set; }
+    }
+}
