+++ src/KPcore/ViewModels/ManageViewModels/VerifyPhoneNumberViewModel.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace KPcore.Models.ManageViewModels
+namespace KPcore.ViewModels.ManageViewModels
 {
     public class VerifyPhoneNumberViewModel
     {
