+++ src/KPcore/ViewModels/ManageViewModels/VerifyTeacherViewModel.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace KPcore.ViewModels.ManageViewModels
+{
+    public class VerifyTeacherViewModel
+    {
+        [Required]
+        public string Code { get; set; }
+    }
+}
