+++ src/KPcore/Views/_ViewImports.cshtml
@@ -1,6 +1,6 @@
 ﻿@using KPcore
 @using KPcore.Models
 @using KPcore.ViewModels.AccountViewModels
-@using KPcore.Models.ManageViewModels
+@using KPcore.ViewModels.ManageViewModels
 @using Microsoft.AspNetCore.Identity
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
